Check that user created widget has expected ID
When creating a widget, while the Ui provides an ID to use, it is
ultimately up to the user to bind this ID to the new widget.
Conceptually, it can happen rather easily that IDs are mixed up. To
inform the user early about such a mishap, this change introduces a
check that verifies that any created widget actually carries the ID that
we have created earlier.