Make event hook handlers to return a single event
In the not-so-distant future all event handling will happen through
async functions and clients will be able to send events through MutCap
objects directly instead of returning a "chain" of them. Hence, once
that day arrives it will no longer be necessary for event hooks to emit
events for communication with other widgets.
In this light, this change adjust the return value of all event hook
functions to be an optional regular event. This event will be returned
from the Ui::handle call, but not passed to widgets in the UI itself.
Because we support multiple event hooks and each can emit an event, we
require that any such events be mergeable. To that end, we introduce the
Mergeable trait. Note that we only require this trait to be implemented
when using MutCap::hook_events, i.e., when installing an event hook.
That means that users not using event hooks do can opt to not make the
event type mergeable.