Introduce post-hook support
The logic for emitting events from an event hook, despite not being very
elegant, enabled an interesting use case given how returned events were
dispatched later on. Specifically, an event hook could emit an event and
rest assured that it would only be delivered once the actual hooked
event has been processed fully. One way to use this capability is to
send the widget for which a hook was installed an event from the hook
itself -- which effectively installs a "callback" that will be triggered
once the event has been processed by other widgets.
There are clients out there that make use of this functionality and
given how cumbersome it is to achieve the same outcome through other
means, this change proposed a replacement mechanism: the notion of a
post-hook, i.e., one that is invoked after the actual event being passed
in has been processed.
To that end, the event hook signature is adjusted to accept an optional
event. If the event is set, we are on the pre-hook path (before the
event has been delivered to non-hooking widgets), if it is not set, the
handler got invoked from the post-hook path.