Make event hooks async
While we no longer allow for returning one or more events from an event
hook handler, it is still considered crucial for clients to be able to
send messages.
This change enables this sending by making event hook handlers async
functions. Hence, it is now possible to use the send and call methods on
the provided capability object from this context.