Remove UiEvent::Directed variant
The message delivery mechanism that we introduced earlier is meant to
act as a complete replacement of the custom event logic that we have
used so far.
As a first step to getting rid of it, this change removes the
UiEvent::Directed variant, that was backing the directed delivery of
events.