Remove UiEvent::Returnable variant
Now that we have support for another communication mechanism that does
not require a transfer of ownership of the message itself, we no longer
have any use for the UiEvent::Returnable variant.
To that end, this change removes this variant and the code supporting
it, which includes the handle_custom_ref handler in the Handleable
trait.