Refrain from using custom event types in event tests
In the future we want to get rid of custom events altogether. However,
we still have a couple of tests that use them for purposes other than
testing custom event types themselves.
With this change we adjust the tests to use regular events instead. This
way, once we want to remove custom event types we will no longer have to
adjust existing test cases.