Remove custom event handling logic
This change finally removes the custom event handling logic. In
particular, we no longer need the Custom variant in the UiEvent and
UnhandledEvent types. Similarly, the handle_custom handler in the
Handleable trait is no longer of use to us. Along with that goes a
simplification of the Ui's internals and a removal of tests that no
longer fulfill any purpose.