Remove no longer required event comparison logic
With the custom event removed the main reason for having custom equality
checking functions is gone: custom events had an Any type in them and
that prevented us from deriving PartialEq for the various event enum
types.
This change removes the different compare_* functions that we introduced
to work around this problem and instead relies on compiler generated
equality checks.