Return only a single event from various event handlers
So far our stance has been that any event being handled could result in
a multitude of new events being emitted. This behavior was mostly driven
by our need to perform multiple operations in response to an event, but
with the new async based message delivery mechanism, we no longer need
to rely on such a kludge to get the desired behavior.
Thus, with this change we adjust the signature of all event handling
functions/methods to only return a single unhandled event. In so doing
we remove one of the main sources for allocation from the event handling
path, while also simplifying the event handling logic inside the Ui
itself.