Remove UnhandledEvent type
We originally introduced the UnhandledEvent type to reflect the
asymmetry between the event send path (where we had undirect and
directed events) and the event return path (with only undirected
events).
Now that we got rid of custom events altogether, the UnhandledEvent type
has strictly the same variants as the UiEvent enum. To that end, this
change removes the UnhandledEvent enum, using UiEvent as a drop-in
replacement.