Remove UiEvent type
This change removes the UiEvent type that we have used for so long. By
now, mostly due to the removal of custom events types in favor of
messages, the type is nothing more than a way for the Ui to add logic
for natively supporting a graceful exit. That, however, is pretty much a
policy that may or may not be required/desired and can equally well be
implemented on the client side. As such, it should not be included in
the crate itself. Without the Quit variant the UiEvent enum is just a
wrapper around the generic event type that the user specified, rendering
it useless, too.