Prevent implementation of Cap trait by users
We cannot think of a good reason for allowing users to provide a custom
implementation of the Cap and MutCap traits and our design simply is not
meant to take custom implementations into consideration. As such, with
this change we seal the Cap trait (and by transitivity MutCap), to
disallow implementation for custom types programmatically.