Allow unknown clippy lints in CI runs
When running clippy as part of the CI pipeline we indirectly may lint
other crates as well. What we have seen happen recently is a dependent
crate using a clippy specific lint that is not yet or no longer known to
the version of clippy that we use. Originally we had hoped that this
case would be covered by the allowing the unknown_lints lint, but it
turns out it actually requires a clippy specific one,
clippy::unknown_clippy_lints.
With this change we explicitly allow this lint as well in our CI
configuration.