Remove children from Placeholder
The Placeholder struct still contained a vector of children, but it has
not been used since the decision to store widget data outside the widget
itself.
To that end, this change removes the member from the type and updates
comments to provide a more up-to-date description of what is happening.