Add links to referenced types to documentation
With a recent version of Rust documentation can now more easily
include references to types mentioned. With this change we make use of
this capability to add explicit links that allow the user to more easily
navigate the crate's documentation.
While at it, we also add a CI step that verifies that we can build the
documentation.