Implement react and respond in Placeholder
The Placeholder is meant to be exactly what the name implies -- a dummy
that is not actually called into. To make sure that we get informed
about violations of this contract, we made its Handleable::handle
implementation panic.
Since then the Handleable trait has gotten support for more methods.
With this change we implement these methods as well, making them panic
in the same way.