Preserve child ordering when showing and hiding widgets
When we devised the widget hiding and showing functionality, we made
each visibility change be directly reflect in a widget's position in its
parent's list of children. Basically, when a widget was being hidden it
was moved all the way to the back of the array (also representing the
lowest z-order). The focused widget, if any, would always occupy the
first index and everything between was used for visible widgets. The
rationale, at the time, was that we could stop rendering once we found
the first widget that was not visible, knowing that all remaining ones
were invisible as well. However, we neither documented (or otherwise
communicated) this property nor did we rely on it.
In fact, it is arguably counter intuitive behavior: making a widget
invisible and then visible again should preserve its order, but that was
not necessarily the case.
This change fixes this issue by not changing the order of children when
hiding or showing. In so doing we simplify the logic a bit. We keep the
behavior focusing, because that's how UIs typically behave.