Address clippy reported issues
In recent versions of Rust clippy suggests a minimization of some
boolean condition that we have. The result is found to be less
expressive. Hence, this change explicitly allows the lint for the
function being flagged.
Furthermore, it complains about our reliance on deref into a super
trait. Our trait explicitly implements Deref into the super trait, so
this is probably a false positive. In any event, it's "only" test code.
Allow the lint explicitly for the time being.