Introduce Renderer::render_done() method
This change introduces another method to the Renderer and Renderable
traits: render_done. This method is invoked when the rendering for the
provided widget is concluded, i.e., when the widget itself as well as
its children have been rendered, at least formally. It is still possible
to perform additional rendering as part of this invocation, which may
come in handy for constructs such as overlays.