Pass correct BBox to Renderer::render_done()
With the introduction of the Renderer::render_done() method, we were
unintentionally passing the bounding box that is meant to be used for
the children to this call as well.
This change fixes this problem by passing the BBox object representing
the outer area to the Renderer::render_done() call instead.