Stop rendering widgets when BBox is empty
When a widget reports an empty BBox to use for its children, there
really is no point in rendering said children -- there is no space for
them anyway.
Hence, with this change we omit calling into the children's rendering
logic altogether if the current bounding box has a dimension that is
zero.