Introduce generic helper for implementing Debug trait
Some standard types unfortunately don't provide a Debug implementation,
making it cumbersome to embed them in a type that is supposed to
auto-derive a Debug impl. Currently this affects the EventHookFn type
and we introduce the internal EventHook handler to circumvent this
problem.
Instead of having such a specialized type, introduce the D struct as a
generic wrapper providing a Debug implementation for the pointer to the
wrapped object.