Fix event handler based hook function installation
In real world scenarios, it is not uncommon to install an event hook
handler function from within an existing event handler. In such a
scenario, the newly installed hook function would still be invoked on
the post-hook path (i.e., without an event being passed in), despite
never having been invoked with the actual event (pre-hook). This
behavior seems confusing and counter-intuitive.
With this change we adjust the logic to not emit a post-hook in such a
scenario. Note that while strictly speaking an incompatible change, we
treat that as a bug fix, as nobody should have relied on this behavior
in the first place.