Store event handlers in Rc objects in TestWidgetData
The TestWidget currently uses Option::take() on each handler function to
decouple itself from the MutCap object that this data is associated
with. That is problematic once we attempt to support/test recursive
invocation of event handlers.
To fix this shortcoming, use an Rc for storing each handler function and
then clone this pointer to achieve the same decoupling, but without
invalidation of the stored handler.