Pass widget Id to TestWidget react & respond handlers
Adjust the TestWidget to to provide the Id of the widget to both the
Handleable::react() and Handleable::respond() handlers. This mimics what
we already do with the callback used to implement Handleable::handle(),
which also receives the widget Id. Having this Id available will provide
us with more flexibility as to what we can do from these handlers.