Introduce MutCap::rehandle() functionality
Introduce a way for recirculating events, retriggering event handling
starting with a specified widget. This can be useful=, for example, in
cases where one widget has the input focus but decided to give it up in
response to an event, but at the same time wanting the newly focused
widget to receive and handle said event, instead of just discarding it.